Ugly fix due to source changes
Code used to search for team name and league name. Currently, the
search is now broken so this is a quick fix to allow the code to
still work.

If you enter a team name, it will only show scores if that team
has a page on the website. League names must match the website
name (but is case insensitive).

Will try to fix properly when I have time...